% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.features.R
\name{book.of.features-package}
\alias{book.of.features-package}
\title{Book of Features}
\description{
\code{book.of.features} seeks to facilitate design-oriented targets during the feature engineering phase for simple datasets.
}
