% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dims.R
\name{create_dims}
\alias{create_dims}
\title{Create Dimension Specs}
\usage{
create_dims(n, d)
}
\arguments{
\item{n}{The total number of values}

\item{d}{The number of dimensions}
}
\value{
A \code{d}-column row-ordered matrix of possible dimensions or (invisibly) a message if no dimensions can be found.
}
\description{
Create Dimension Specs
}
\examples{
k <- create_dims(n = 20, d = 3)
# Create an array with dimensions sampled from `k`:
array(sample(x = 20, size = prod(k[1, ]), replace = TRUE), dim = k[sample(nrow(k), 1), ])

k <- create_dims(n = 56, d = 4)
# Create an array with dimensions sampled from `k`:
array(sample(x = 20, size = prod(k[1, ]), replace = TRUE), dim = k[sample(nrow(k), 1), ])

}
