% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.features.R
\name{define}
\alias{define}
\title{Define a Data Operation}
\usage{
define(x, ...)
}
\arguments{
\item{x}{The input data}

\item{...}{(\code{\link[rlang]{dots_list}}): Formulae for which the left-hand side (LHS) is an expression containing the operation, and the right-hand side (RHS) contains column names that form a grouping set for the operation (i.e., \code{<expression> ~ col_1 + col_2 + ...}):
\itemize{
\item{If the form \code{<LHS>~ .} is given, the LHS executes using all columns as the grouping set}
\item{If the form \code{<LHS>~ 1} is given, the LHS executes without grouping}
\item{If no LHS is given, the operation defaults to selection based on the RHS}
}}
}
\value{
The data modified
}
\description{
\code{define} allows one to operate on data using one or more formula-based definitions.
}
