% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.features.R
\name{join.mapper}
\alias{join.mapper}
\title{Create a Data Join Map}
\usage{
join.mapper(
  map_name = "new_join_map",
  obj_names,
  field_names = "*",
  env = parent.frame(),
  clean = FALSE
)
}
\arguments{
\item{map_name}{The output map name to use in assignment}

\item{obj_names}{(string[]) The names of objects to join}

\item{field_names}{(string[]) One or more strings and REGEX patterns used to define the field names to use for possible joins: matching is either REGEX or identity, and support for aliasing via \code{"primary_col==alias_col"} (note the quotes) is supported.}

\item{env}{The environment where source objects are found; this is also the environment of the assigned output.}

\item{clean}{(logical) \code{TRUE} indicates that the object should be removed from \code{env} before creating the output}
}
\value{
If a \code{\link[DBOE]{DBOE}} DSN environment is passed to \code{env}, a local environment is used to contain the output of special code that transforms \code{env$metamap} into environment objects that can be used for creating the map: this should be stored or piped into \code{\link{join.reduce}}; otherwise, a \code{\link[data.table]{data.table}} object used join datasets via \code{\link{join.reduce}}
}
\description{
\code{join.mapper} Creates a map with which datasets can be joined using \code{\link[data.table]{data.table}} methods.  Cross-environment objects are not supported, but objects in attached environment \emph{might} work.  Also, the output is given attribute \code{"env"} to store the value of \code{env}
}
