% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.features.R
\name{join.reduce}
\alias{join.reduce}
\title{Join-Reduce Multiple Datasets}
\usage{
join.reduce(
  jmap,
  x.names,
  i.names,
  filters = TRUE,
  dt.key,
  env,
  clean = FALSE,
  .debug = FALSE
)
}
\arguments{
\item{jmap}{The join-map (see \code{link{join.mapper}})}

\item{x.names}{(string[]) Names or REGEX patterns indicating the outer table: multiple values will be concatenated into a delimited string}

\item{i.names}{(string[]) Names or REGEX patterns indicating the inner table to be joined: multiple values will be concatenated into a delimited string}

\item{filters}{(expression[[]]) \code{data.table}-friendly expression to limit the rows of the output. A length-2 list indicates \emph{row} and \emph{column} expressions respectively.}

\item{dt.key}{(string[], symbol[]) Names that participate in creating the \code{\link[data.table]{key}} for the output}

\item{env}{The environment in which the output should be assigned.  If empty, it defaults to \code{attr(jmap, "env")}}

\item{clean}{(logical) \code{TRUE} indicates that the object should be removed from \code{env} before creating the output}

\item{.debug}{(logical) Debugging flag: no output is produced}
}
\value{
Combined datasets assigned to the designated environment
}
\description{
\code{join.reduce} leverages \code{\link[data.table]{data.table}} fast joins and \code{\link[purrr]{reduce}} from package \code{purr}
}
