% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.date_time.R
\name{make.date_time}
\alias{make.date_time}
\title{Date-Time Sequence Generator}
\usage{
make.date_time(
  add_vec = 1:7,
  var_start = Sys.Date(),
  var_form = "\%Y-\%m-\%d 00:00:00",
  var_tz = "",
  var_interval = "days"
)
}
\arguments{
\item{add_vec}{(numeric): Vector of integer values to add}

\item{var_start}{(datetime): Vector of integer values used to generate datetime results}

\item{var_form}{(string): Format for the datetime values}

\item{var_tz}{(string): Timezone for the datetime values}

\item{var_interval}{(string): The unit of time for \code{add_vec}}
}
\value{
A temporal vector
}
\description{
\code{make.date_time} serves as a wrapper for \code{\link[stringi]{stri_datetime_add}}
}
\seealso{
Other Data Generation: 
\code{\link{formulate}()}
}
\concept{Data Generation}
