% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.features.R
\name{make.islands}
\alias{make.islands}
\title{make.islands}
\usage{
make.islands(
  srcData,
  mapFields,
  timeFields,
  timeout = GAP > timeout,
  boundaryName = "episode",
  archipelago = TRUE,
  show.all = FALSE,
  debug = FALSE
)
}
\description{
An alias for \code{\link{continuity}}
}
\seealso{
Other DEPRECATED: 
\code{\link{xform.basis_vector}()},
\code{\link{xform.sigmoid}()}
}
\concept{DEPRECATED}
