% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.features.R
\name{make.windows}
\alias{make.windows}
\title{Serial Window Maker}
\usage{
make.windows(
  series,
  window.size,
  increment = 1,
  post = eval,
  debug = FALSE,
  ...
)
}
\arguments{
\item{series}{A list or vector object from which incremental subsets (windows) of a fixed size are chosen}

\item{window.size}{(integer) The size of the subset (window) to select}

\item{increment}{(integer) The number of elements by which iteration should advance}

\item{post}{(function) A post-processing function on the return object having class "data.table" and single column "window"}

\item{debug}{(logical | FALSE) When \code{TRUE}, additional information is printed to console for debugging purposes

Sets are the result of forward-moving partitioning:
\enumerate{
  \item \code{window.size}: The size of each partition (W)
  \item \code{increment}: The number of items to increment before selecting the next W items (W + i)
}}
}
\value{
A serialized collection-list of partitions (windows), each window containing a subset of size \code{window.size}
}
\description{
\code{make.windows} takes as its first argument a series of values and produces a list of partitioned sets referred to as a "window"
}
\seealso{
Other Data Generation: 
\code{\link{bin.windows}()},
\code{\link{continuity}()},
\code{\link{make.date_time}()}
}
\concept{Data Generation}
