% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.features.R
\name{sigmoid}
\alias{sigmoid}
\title{{0,N} Sigmoid Scaler}
\usage{
sigmoid(input, family = "logistic", center = mean, debug = FALSE, ...)
}
\arguments{
\item{input}{(numeric vector or coercible vector)}

\item{family}{(string | "logistic") The family of sigmoid equation to use: Currently, only "logistic", "generalized", and "gompertz" are supported}

\item{center}{A function that returns the 'center' of \code{input}}

\item{...}{Valid elements that populate the arguments of \code{type} based on the selected \code{family}}
}
\value{
A numeric vector of domain {0, N}
}
\description{
\code{sigmoid} Scales the input to a range of {0,N} using the sigmoid function
}
\details{
The input must contain positive and negative values when \code{centered == FALSE}

Function arguments for each sigmoid family are as follows:
\enumerate{
  \item \code{"generalized"}: \code{list(A, K, C, Q, B, v)}
  \item \code{"logistic"}: \code{list(L, K)}
  \item \code{"gompertz"}: \code{list(A, B, C)}
}
Families \code{tanh}, \code{atan}, and \code{guder}(mannian) only take the input as its argument
}
\section{References}{

\enumerate{
\item{\href{https://en.m.wikipedia.org/wiki/Generalised_logistic_function}{Generalized Logistic Function}}
\item{\href{https://en.m.wikipedia.org/wiki/Sigmoid_function}{Sigmoid Function}}
\item{\href{https://en.m.wikipedia.org/wiki/Gompertz_function}{Gompertz Function}}
\item{\href{https://en.m.wikipedia.org/wiki/Gudermannian_function}{Gudermannian Function}}
}
}

