% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.features.R
\name{xform.basis_vector}
\alias{xform.basis_vector}
\title{xform.basis_vector}
\usage{
xform.basis_vector(
  fvec,
  avec = rep(1, length(fvec)),
  bvec = sort(unique(fvec)),
  cmp_test = `==`,
  logical.out = FALSE,
  regex = FALSE,
  chatty = FALSE
)
}
\description{
An alias for \code{\link{logic_map}}
}
\seealso{
Other DEPRECATED: 
\code{\link{make.islands}()},
\code{\link{xform.sigmoid}()}
}
\concept{DEPRECATED}
