% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.features.R
\name{bin.windows}
\alias{bin.windows}
\title{Create Bins From Integer Factor}
\usage{
bin.windows(i = 1, use.bin = NULL, as.factor = FALSE, ...)
}
\arguments{
\item{i}{(integer[]) An integer scalar, vector, or n-dimensional object executed conditionally as follows:
\itemize{
\item if a vector of length = 1, a zero-based sequence up to \code{abs(i) } is used
\item if a vector of length = 2, a sequence is created from the values in the order given
\item if a vector of length >= 3, the raw values
\item if n-dimensional, recursion along the last dimension given by \code{dim(i)} until a vector is detected
}}

\item{use.bin}{The bin size to use: when empty, the smallest prime number or integer factor in \code{i} within the range of \code{i} is used.}

\item{as.factor}{(logical) Should the output be converted into a factor?}

\item{...}{(not used)}
}
\value{
A character (or factor) vector the length of the input, as "binned" representations.  If the input is dimensional, an array of the same dimensions is returned
}
\description{
\code{bin.windows} creates binned ranges based on the minimum factor of the integer input (\code{i}) or user-supplied value.
}
\seealso{
Other Data Generation: 
\code{\link{make.date_time}()},
\code{\link{make.windows}()}
}
\concept{Data Generation}
