% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.features.R
\name{bin.windows}
\alias{bin.windows}
\title{Create Bins From Integer Factor}
\usage{
bin.windows(i = 1, use.bin = NULL, min.factor = 1)
}
\arguments{
\item{i}{(integer[[[]]]) An integer scalar, vector, or n-dimensional object executed conditionally as follows:
\itemize{
\item if a vector of length = 1, a zero-based sequence up to \code{abs(i) } is used
\item if a vector of length = 2, a sequence is created from the values in the order given
\item if a vector of length >= 3, the raw values
\item if n-dimensional, recursion along the last dimension given by `dim(i)` until a vector is detected
}}

\item{use.bin}{(integer) The bin size to use: should be greater than zero (0), overrides the internal effects of argument `min.factor`.}

\item{min.factor}{(integer) The minimum factor of of `i` allowed when `use.bin` is less than or equal to one (1).  This becomes the bin size.}
}
\value{
An ordered \code{\link[base]{factor}} with levels defined on the bin size.  If the input is dimensional, an array of the same dimensions is returned
}
\description{
`bin.windows()` creates binned ranges based on the minimum factor of the integer input (`i`) or user-supplied value.
}
