% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulate.R
\name{formulate}
\alias{formulate}
\title{Make A Formula}
\usage{
formulate(lhs = NULL, ...)
}
\arguments{
\item{lhs}{(string|symbol) The left-hand side of the formula}

\item{...}{(dots-list) Elements to use on the right-hand side of the formula}
}
\value{
A formula
}
\description{
\code{formulate} takes the inputs and assembles a \code{\link[stats]{formula}} object
}
\section{\code{...}}{

\itemize{
\item{Length-1 elements are treated as terms}
\item{Vectors are reduced as individual terms}
\item{Refer to \code{\link[stats]{formula}} for valid syntax}
}
}

\examples{
formulate(score)
formulate(score, )
formulate(score, -that)
formulate(score, this)
formulate(score, this:that)
formulate(score, this*that)
formulate(score, this, -that, c(that:other, high*low))

}
\seealso{
Other Data Generation: 
\code{\link{bin.windows}()},
\code{\link{make.date_time}()},
\code{\link{make.windows}()}
}
\concept{Data Generation}
