% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.features.R
\name{make.quantiles}
\alias{make.quantiles}
\title{Quantiles Transformation}
\usage{
make.quantiles(x, as.factor = FALSE, ...)
}
\arguments{
\item{x}{The input vector}

\item{as.factor}{(logical) Should the output be returned as a factor?}

\item{...}{(\code{\link[rlang]{dots_list}}): Additional arguments sent to \code{\link[stats]{quantile}}.  Defaults are as follows:\cr
\itemize{
\item{\code{probs = seq(0, 1, 0.25)}}
\item{\code{na.rm = FALSE}}
\item{\code{names = TRUE}}
\item{\code{type = 7}}
\item{\code{digits = 7}}
\item{\code{...}}
}}
}
\value{
A quantile representation of the input
}
\description{
\code{make.quantiles} is a wrapper for \code{\link[stats]{quantile}} replacing the input with calculated values.  Unmatched quantiles will return \code{NA} in order to match the length of the input vector.
}
\note{
Any indeterminate probabilities relative to the input will return \code{NA}
}
\examples{
sample(900, 20) |> (\(x) data.frame(x = x, q = book.of.features::make.quantiles(x, seq(0, 1, .1), as.factor = TRUE)))()
sample(900, 20) |> (\(x) data.frame(x = x, q = book.of.features::make.quantiles(x, seq(0, 1, .1), as.factor = FALSE)))()

}
