% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book.of.features.R
\name{xform.sigmoid}
\alias{xform.sigmoid}
\title{xform.sigmoid}
\usage{
xform.sigmoid(input, family = "logistic", center = mean, debug = FALSE, ...)
}
\description{
An alias for \code{\link{sigmoid}}
}
\seealso{
Other DEPRECATED: 
\code{\link{make.islands}()},
\code{\link{xform.basis_vector}()}
}
\concept{DEPRECATED}
